requirements.txt
airtest==1.0.27
altgraph==0.16.1
certifi==2019.6.16
chardet==3.0.4
Click==7.0
comtypes==1.1.7
configparser==3.8.1
decorator==4.4.0
facebook-wda==0.3.6
func-timeout==4.3.5
future==0.17.1
hrpc==1.0.8
idna==2.8
Jinja2==2.10.1
MarkupSafe==1.1.1
mss==4.0.3
numpy==1.15.1
opencv-contrib-python==3.4.2.17
pefile==2019.4.18
Pillow==6.1.0
pocoui==1.0.76
py==1.8.0
PyInstaller==3.5
pypiwin32==223
PyQt5==5.13.0
PyQt5-sip==4.19.18
pyqt5-tools==5.13.0.1.5
python-dotenv==0.10.3
pywin32==224
pywin32-ctypes==0.2.0
pywinauto==0.6.3
requests==2.22.0
retry==0.9.2
six==1.12.0
urllib3==1.25.3
websocket-client==0.56.0
